move BlockSlice from gen-models to gen-graph
